+++ includes/Core/class-vlt-deactivator.php
@@ -6,6 +6,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 class VLT_Deactivator {
 
 	public static function deactivate() {
+		wp_clear_scheduled_hook( 'vlt_daily_cleanup' );
 		flush_rewrite_rules();
 	}
 }